+++ packages/index/src/write.js
@@ -1,17 +1,20 @@
-import fs from 'fs';
+import path from 'path';
 
-import { writeJsonGzip } from '@home-gallery/common';
+import { writeJsonGzip, promisify } from '@home-gallery/common';
 
 import { byDirDescFileAsc } from './utils.js';
 
-export const writeIndex = (filename, index, cb) => {
-  index.created = new Date().toISOString();
-  index.data.sort(byDirDescFileAsc);
-  const tmp = `${filename}.tmp`;
-  writeJsonGzip(tmp, index, err => {
-    if (err) {
-      cb(err);
-    }
-    fs.rename(tmp, filename, err => cb(err, err ? null : index));
-  });
+const asyncWriteJsonGzip = promisify(writeJsonGzip)
+
+export const writeIndex = async (directory, filename, entries, options) => {
+  const index = {
+    type: 'home-gallery/fileindex@1.0',
+    created: new Date().toISOString(),
+    base: path.resolve(directory),
+    data: entries.sort(byDirDescFileAsc)
+  }
+  if (options.dryRun) {
+    return index
+  }
+  return asyncWriteJsonGzip(filename, index).then(() => index)
 }