+++ packages/server/src/app.js
@@ -64,10 +64,10 @@ export function createApp(context) {
   const app = express();
   app.disable('x-powered-by')
   app.enable('trust proxy')
+  app.use(augmentReqByUserMiddleware())
   app.use(loggerMiddleware())
 
   const router = express.Router()
-  router.use(augmentReqByUserMiddleware())
   router.use(cors());
   router.use(compression({ filter: shouldCompress }))
 
@@ -106,8 +106,7 @@ export function createApp(context) {
     const disabled = config?.webapp?.disabled || []
     const plugins = pluginApi.pluginEntries
     const entries = await getFirstEntries(50, req)
-    const sources = config.sources
-      .filter(source => source.downloadable && !source.offline)
+    const sources = (config.sources || []).filter(source => source.downloadable && !source.offline)
       .map(source => {
         const indexName = path.basename(source.index).replace(/\.[^.]+$/, '')
         return {